requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.0.3
Flask-Cors==3.0.10
langchain==0.0.292
python-dotenv==0.19.2
chromadb==0.4.10
openai==0.28.0
tiktoken==0.3.2
